test: move password confirmation tests to Livewire component test
The original PasswordConfirmationTest.php was testing the password confirmation functionality but was not specific to the Livewire component.

This commit moves the tests to the ConfirmPasswordTest.php file, which is the Livewire component test file. This change improves the organization and clarity of the tests. The PasswordConfirmationTest.php file is removed as it is now redundant.
