+++ tests/Feature/Livewire/Auth/ConfirmPasswordTest.php
@@ -67,4 +67,39 @@ class ConfirmPasswordTest extends TestCase
 
         $this->assertFalse(session()->has('auth.password_confirmed_at'));
     }
+
+    public function test_confirm_password_screen_can_be_rendered(): void
+    {
+        /** @var User */
+        $user = User::factory()->create();
+
+        $response = $this->actingAs($user)->get('/user/confirm-password');
+
+        $response->assertStatus(200);
+    }
+
+    public function test_password_can_be_confirmed(): void
+    {
+        /** @var User */
+        $user = User::factory()->create();
+
+        $response = $this->actingAs($user)->post('/user/confirm-password', [
+            'password' => 'password',
+        ]);
+
+        $response->assertRedirect();
+        $response->assertSessionHasNoErrors();
+    }
+
+    public function test_password_is_not_confirmed_with_invalid_password(): void
+    {
+        /** @var User */
+        $user = User::factory()->create();
+
+        $response = $this->actingAs($user)->post('/user/confirm-password', [
+            'password' => 'wrong-password',
+        ]);
+
+        $response->assertSessionHasErrors();
+    }
 }
